Refactored out storage implementations into their own file
